+++ repositories/store.repository.js
@@ -1,4 +1,5 @@
-const { Store } = require('../models');
+const { Store, Menu } = require('../models');
+const { Op } = require('sequelize');
 
 class StoreRepository {
   getStore = async () => {
@@ -8,9 +9,27 @@ class StoreRepository {
   };
 
   getStoreDetail = async (storeId) => {
-    const oneStoreData = await Store.findAll({ where: { id: storeId } });
+    const oneStoreData = await Store.findOne({ where: { id: storeId } });
 
     return oneStoreData;
+  };
+
+  searchStore = async (searchKeyword) => {
+    const searchStore = await Store.findAll({
+      where: { store_name: { [Op.substring]: searchKeyword } },
+    });
+    return searchStore;
+  };
+  searchMenu = async (searchKeyword) => {
+    const searchMenu = await Menu.findAll({
+      where: {
+        [Op.or]: [
+          { menu: { [Op.substring]: searchKeyword } },
+          { category: { [Op.substring]: searchKeyword } },
+        ],
+      },
+    });
+    return searchMenu;
   };
 }
