+++ routes/store.route.js
@@ -5,7 +5,8 @@ const StoreController = require('../controllers/store.controller');
 const storeController = new StoreController();
 
 router.use('/:storeId/reviews', reviewRouter);
-router.use('/', storeController.getStore);
-router.use('/:storeId', storeController.getStoreDetail);
+router.get('/', storeController.getStore);
+router.get('/:storeId', storeController.getStoreDetail);
+router.post('/search', storeController.search);
 
 module.exports = router;
